Added base materials

+++ Engine/Source/Platform/DirectX11/DirectX11ConstantBuffer.cpp
@@ -13,7 +13,7 @@ namespace Engine
 	{
 		D3D11_BUFFER_DESC bufferDesc;
 		ZeroMemory(&bufferDesc, sizeof(bufferDesc));
-		bufferDesc.ByteWidth = stride;
+		bufferDesc.ByteWidth = (UINT)stride;
 		bufferDesc.Usage = D3D11_USAGE_DYNAMIC;
 		bufferDesc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
 		bufferDesc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
